requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.0
Jinja2==3.1.2
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
python-decouple==3.8
